Separate testing dependencies
