+++ deps.ts
@@ -1,8 +1,3 @@
 export type { ServeInit } from "https://deno.land/std@0.165.0/http/server.ts";
 export { serve } from "https://deno.land/std@0.165.0/http/server.ts";
 export { parse } from "https://deno.land/std@0.165.0/flags/mod.ts";
-
-export {
-  assertEquals,
-  assertRejects,
-} from "https://deno.land/std@0.165.0/testing/asserts.ts";
