+++ test.deps.ts
@@ -1,0 +1,4 @@
+export {
+  assertEquals,
+  assertRejects,
+} from "https://deno.land/std@0.165.0/testing/asserts.ts";
